requirements.txt
asgiref==3.5.2
Django==3.2.15
django-autoslug==1.9.8
django-crispy-forms==1.14.0
django-environ==0.9.0
Pillow==9.3.0
psycopg2==2.9.5
pytz==2022.2.1
sqlparse==0.4.3
typing_extensions==4.3.0
